# BackEnd/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{id}": {
      "put": {
        "operationId": "UserController_update",
        "summary": "Atualizar um usuário",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário atualizado com sucesso"
          }
        },
        "tags": [
          "Usuário"
        ]
      },
      "delete": {
        "operationId": "UserController_delete",
        "summary": "Excluir um usuário",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário excluído com sucesso"
          }
        },
        "tags": [
          "Usuário"
        ]
      }
    },
...
    "/avaliacao/media/{userId}": {
      "get": {
        "operationId": "AvaliacaoController_obterMediaAvaliacoes",
        "summary": "Obter a média de avaliações de um usuário",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Média de avaliações obtida com sucesso"
          }
        },
        "tags": [
          "Avaliação"
        ]
      }
    },
...
    "/hist-caronas/user/{userId}": {
      "get": {
        "operationId": "HistCaronasController_findByUserId",
        "summary": "Obter caronas no histórico por ID do usuário",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de caronas obtido com sucesso"
          }
        },
        "tags": [
          "Histórico de Caronas"
        ]
      }
    },
...
    "/veiculo/{userId}": {
      "get": {
        "operationId": "VeiculoController_findById",
        "summary": "Obter veículos por ID do usuário",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Veículos obtidos com sucesso"
          }
        },
        "tags": [
          "Veículo"
        ]
      }
    },
    "/veiculo/{id}": {
      "delete": {
        "operationId": "VeiculoController_delete",
        "summary": "Excluir um veículo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do veículo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Veículo excluído com sucesso"
          }
        },
        "tags": [
          "Veículo"
        ]
      }
    }
  },
  "info": {
    "title": "BORAaí",
    "description": "Backend BORAaí",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}